feat:  - add point in sirius

+++ app/src/main/java/com/habit/tracker/data/source/remote/model/dto/GeoDTO.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 class GeoDTO(
     @SerializedName("long")
-    val longitude: Double,
+    val longitude: Float,
     @SerializedName("lat")
-    val latitude: Double,
+    val latitude: Float,
 )
