+++ app/src/main/java/com/habit/tracker/presentation/view/MapFragment.kt
@@ -144,7 +144,7 @@ class MapFragment : Fragment() {
                     })
                     val pointAnnotationOptions: PointAnnotationOptions =
                         PointAnnotationOptions()
-                            .withPoint(Point.fromLngLat(geo.long, geo.lat))
+                            .withPoint(Point.fromLngLat(geo.long.toDouble(), geo.lat.toDouble()))
                             .withIconImage(b)
                     pointAnnotationManager.create(pointAnnotationOptions)
                 }
